+++ test_diffusion_kernel.py
@@ -10,17 +10,17 @@ class TestDiffusionKernel(object):
         self.func = self.mod.get_function("temperature_update16x16")
 
     def test_prepared_call(self):        
-        f = np.ones([3, 3, 3], dtype=np.float32)
-        f = f*np.array([1,2,3], dtype=np.float32)
+        f = np.ones([3, 3, 3], dtype=np.float64)
+        f = f*np.array([1,2,3], dtype=np.float64)
         f_gpu = gpuarray.to_gpu(f)
 
-        dfdx = np.zeros([3, 3, 3], dtype=np.float32)
+        dfdx = np.zeros([3, 3, 3], dtype=np.float64)
         dfdx_gpu = gpuarray.to_gpu(dfdx)
         
         # run the kernel:
-        self.func.prepare([np.intp, np.intp, np.float32, np.float32,
+        self.func.prepare([np.intp, np.intp, np.float64, np.float64,
                              np.intc, np.intc, np.intc, 
-                                 np.float32, np.float32, np.float32])
+                                 np.float64, np.float64, np.float64])
 
         self.func.prepared_call((1,1,1), (16, 16, 1), 
                            f_gpu.gpudata, dfdx_gpu.gpudata, 1., 1., 3, 3, 3, 1, 1, 1)
